# apps/next/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Datasets Generator API",
    "description": "API to generate datasets from web scraping",
    "version": "0.0.1"
  },
  "paths": {
...
    "/dataset/generate/file": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "Generate a dataset from an uploaded file (PDF or image)",
        "description": "Create or extend a dataset from an uploaded PDF or image. Each page is transcribed with a vision model (the vision role default unless model_vlm is given), then mined for question-answer pairs.",
        "operationId": "create_dataset_for_file_dataset_generate_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_dataset_for_file_dataset_generate_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Dataset created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetGenerationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters or source",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dataset/generate/url": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "Generate a dataset from a web page",
        "description": "Create or extend a dataset from one web page (no crawling): the page is fetched, reduced to text, cleaned, then mined for question-answer pairs. Private/loopback addresses are refused.",
        "operationId": "create_dataset_for_url_dataset_generate_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UrlGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Dataset created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetGenerationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters or source",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dataset": {
      "post": {
        "tags": [
          "dataset"
        ],
        "summary": "Create Dataset",
        "description": "Create a new (empty) dataset of yours.",
        "operationId": "create_dataset_dataset_post",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Name of the new dataset",
              "title": "Name"
            },
            "description": "Name of the new dataset"
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Optional dataset description",
              "title": "Description"
            },
            "description": "Optional dataset description"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "dataset"
        ],
        "summary": "Get All Datasets",
        "description": "Retrieve all your datasets, or a specific one (by name).",
        "operationId": "get_all_datasets_dataset_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Optional dataset name to get a specific dataset's details",
              "title": "Dataset Id"
            },
            "description": "Optional dataset name to get a specific dataset's details"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DatasetResponse"
                    },
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DatasetResponse"
                      }
                    }
                  ],
                  "title": "Response Get All Datasets Dataset Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_dataset_for_file_dataset_generate_file_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "PDF or image file to mine for Q&A"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name"
          },
          "target_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Language"
          },
          "model_qa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Qa",
            "description": "Model reference for QA"
          },
          "model_vlm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Vlm",
            "description": "Model reference for page transcription"
          },
          "similarity_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Similarity Threshold",
            "default": 0.85
          },
          "persist": {
            "type": "boolean",
            "title": "Persist",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "file",
          "dataset_name"
        ],
        "title": "Body_create_dataset_for_file_dataset_generate_file_post"
      },
...
      "DatasetGenerationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "ID of the dataset"
          },
          "qa_pairs": {
            "items": {
              "$ref": "#/components/schemas/QAPair"
            },
            "type": "array",
            "title": "Qa Pairs",
            "description": "List of generated question-answer pairs"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "description": "Name of the dataset"
          },
          "model_cleaning": {
            "type": "string",
            "title": "Model Cleaning",
            "description": "Model used to clean (URL) or transcribe (file) the source"
          },
          "target_language": {
            "type": "string",
            "title": "Target Language",
            "description": "Target language used"
          },
          "model_qa": {
            "type": "string",
            "title": "Model Qa",
            "description": "Model used for QA generation"
          },
          "similarity_threshold": {
            "type": "number",
            "title": "Similarity Threshold",
            "description": "Similarity threshold used"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions",
            "description": "Total number of generated questions"
          },
          "pages_crawled": {
            "type": "integer",
            "title": "Pages Crawled",
            "description": "Number of pages fetched and processed",
            "default": 1
          },
          "processing_time": {
            "type": "number",
            "title": "Processing Time",
            "description": "Processing time in seconds"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PipelineStep"
            },
            "type": "array",
            "title": "Steps",
            "description": "Timeline of the pipeline steps"
          },
          "scraped_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scraped Content",
            "description": "Text extracted from the source, as fed to QA generation"
          },
          "persisted": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persisted",
            "description": "Summary of the stored version (dataset name, version, run) when the pairs were saved at generation time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "qa_pairs",
          "dataset_name",
          "model_cleaning",
          "target_language",
          "model_qa",
          "similarity_threshold",
          "total_questions",
          "processing_time"
        ],
        "title": "DatasetGenerationResponse",
        "description": "Model for dataset generation response",
        "example": {
          "dataset_name": "my_dataset",
          "id": "123e4567-e89b-12d3-a456-426614174000",
          "model_cleaning": "gpt-3.5-turbo",
          "model_qa": "gpt-4",
          "processing_time": 45.2,
          "qa_pairs": [
            {
              "answer": "The document discusses machine learning techniques.",
              "question": "What is the main topic of this document?"
            }
          ],
          "similarity_threshold": 0.85,
          "target_language": "en",
          "total_questions": 50
        }
      },
      "DatasetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "target_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Language"
          },
          "qa_sources_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qa Sources Count"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "DatasetResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Detailed error description"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code",
            "description": "Specific error code"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Model for error responses",
        "example": {
          "detail": "Model 'invalid-model' not in available models: ['gpt-3.5-turbo', 'gpt-4']",
          "error_code": "INVALID_MODEL"
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PipelineStep": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Stable identifier of the step"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "Human-readable step name"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Outcome of the step: success, warning or error"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms",
            "description": "Step duration in milliseconds"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Log line summarising what happened",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "status",
          "duration_ms"
        ],
        "title": "PipelineStep",
        "description": "One step of the generation pipeline, for the frontend timeline."
      },
...
      "QAPair": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question",
            "description": "Generated question"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "description": "Corresponding answer"
          }
        },
        "type": "object",
        "required": [
          "question",
          "answer"
        ],
        "title": "QAPair",
        "description": "Model for a question-answer pair"
      },
...
      "UrlGenerationRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "http(s) URL of the page to mine"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "description": "Name of the dataset to create or extend"
          },
          "target_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Language",
            "description": "Target language for QA generation"
          },
          "model_cleaning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Cleaning",
            "description": "Model reference for text cleaning, e.g. \"openai:gpt-4o-mini\" (default: the cleaning role default)"
          },
          "model_qa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Qa",
            "description": "Model reference for QA generation, e.g. \"claude:claude-sonnet-5\" (default: the qa role default)"
          },
          "similarity_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Similarity Threshold",
            "description": "Question similarity at or above which a pair is a duplicate (0.0-1.0): embedding cosine across the dataset when the local model is available, else a lexical ratio within the same source",
            "default": 0.85
          },
          "persist": {
            "type": "boolean",
            "title": "Persist",
            "description": "Store the generated pairs and record a version at generation time",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "url",
          "dataset_name"
        ],
        "title": "UrlGenerationRequest",
        "description": "Generate a dataset from one web page (no crawling)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}